fix: moving files to features folder
